services/dashboards/public/requirements.txt
# Copy executive requirements
dash==2.14.2
dash-bootstrap-components==1.5.0
plotly==5.18.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
pandas==2.1.4
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3
python-json-logger==2.0.7
gunicorn==21.2.0
